Removed repaedted code and add it into common place 🎨
